requirements.txt
# ── Cloud backend (Render / Railway / Fly.io) ──────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
pandas>=2.2.0
openpyxl>=3.1.0          # xlsx support
python-dotenv>=1.0.0
python-multipart>=0.0.9  # file uploads

# ── Edge node (local / Raspberry Pi) ──────────────────────────────────────────
ultralytics>=8.2.0        # YOLOv8
opencv-python>=4.9.0      # cv2
numpy>=1.26.0
websockets>=12.0
